# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |-
    Connect with your friends effortlessly using WASAText! Send and receive messages, 
    whether one-on-one or in groups, all from the convenience of your PC. 
    Enjoy seamless conversations with text or GIFs and easily stay in touch 
    through your private chats or group discussions.
  version: "1.0.0"
...
servers:
  - url: "http://localhost:3000"
    description: Development server
paths:
  /session:
    post:
      tags: ["Authentication"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: User login credentials
              properties:
                name:
                  type: string
                  description: User's display name for login or registration
                  example: Maria
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: Login response with user identifier
                properties:
                  identifier:
                    type: string
                    description: Unique user identifier to use as authentication token
                    pattern: '^[a-zA-Z0-9]+$'
                    minLength: 1
                    maxLength: 50
                    example: "abcdef012345"
                required:
                  - identifier
              example:
                identifier: "user123abc456"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /groups:
    post:
      tags: ["Groups"]
      summary: Create a new group
      description: Create a new group conversation with other users
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        description: Group details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for creating a new group conversation
              properties:
                name:
                  type: string
                  description: Display name for the new group
                  pattern: '^.+$'
                  minLength: 1
                  maxLength: 50
                  example: "Family Chat"
                memberIds:
                  type: array
                  description: List of user IDs to add to the group
                  minItems: 1
                  maxItems: 99
                  items:
                    type: string
                    description: User identifier
                    pattern: '^[a-zA-Z0-9]+$'
                    minLength: 1
                    maxLength: 50
                  example: ["user123", "user456"]
              required:
                - name
                - memberIds
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                id: "group123"
                name: "Family Chat"
                photoUrl: "https://api.example.com/groups/group123/photo"
                members:
                  - id: "user123"
                    name: "Maria"
                    photoUrl: "https://api.example.com/users/user123/photo"
                  - id: "user456"
                    name: "Marco"
                    photoUrl: "https://api.example.com/users/user456/photo"
                createdAt: "2023-12-07T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/groupId'
    put:
      tags: ["Groups"]
      summary: Update group name
      description: Update the name of a group
      operationId: setGroupName
      security:
        - BearerAuth: []
      requestBody:
        description: New group name
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for updating group name
              properties:
                name:
                  type: string
                  description: New display name for the group
                  pattern: '^.+$'
                  minLength: 1
                  maxLength: 50
              required:
                - name
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 100
                    example: "Group name updated successfully"
              example:
                message: "Group name updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/{groupId}/photo:
    parameters:
      - $ref: '#/components/parameters/groupId'
    put:
      tags: ["Groups"]
      summary: Set group photo
      description: Upload or update the group's photo
      operationId: setGroupPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: Group photo
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: PNG image file data
              minLength: 1
              maxLength: 10485760
          image/jpeg:
            schema:
              type: string
              format: binary
              description: JPEG image file data
              minLength: 1
              maxLength: 10485760
          image/gif:
            schema:
              type: string
              format: binary
              description: GIF image file data
              minLength: 1
              maxLength: 10485760
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 100
                    example: "Group photo updated successfully"
              example:
                message: "Group photo updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: Photo too large
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: ["Groups"]
      summary: Get group photo
      description: Retrieve the group's photo
      operationId: getGroupPhoto
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Photo retrieved successfully
          content:
            image/png:
              schema:
                type: string
                format: binary
                description: PNG image file data
                minLength: 1
                maxLength: 10485760
            image/jpeg:
              schema:
                type: string
                format: binary
                description: JPEG image file data
                minLength: 1
                maxLength: 10485760
            image/gif:
              schema:
                type: string
                format: binary
                description: GIF image file data
                minLength: 1
                maxLength: 10485760
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Group not found or no photo available
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/groupId'
    get:
      tags: ["Groups"]
      summary: Get group members
      description: Get list of all members in a group
      operationId: getGroupMembers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Group members retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: Response containing group members list
                properties:
                  members:
                    type: array
                    description: List of group members
                    minItems: 0
                    maxItems: 100
                    items:
                      $ref: '#/components/schemas/User'
                required:
                  - members
              example:
                members:
                  - id: "user123"
                    name: "Maria"
                    photoUrl: "https://api.example.com/users/user123/photo"
                  - id: "user456"
                    name: "Marco"
                    photoUrl: "https://api.example.com/users/user456/photo"
                  - id: "user789"
                    name: "Anna"
                    photoUrl: "https://api.example.com/users/user789/photo"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: ["Groups"]
      summary: Add user to group
      description: Add a user to an existing group
      operationId: addToGroup
      security:
        - BearerAuth: []
      requestBody:
        description: User to add to group
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for adding a user to the group
              properties:
                userId:
                  type: string
                  description: ID of the user to add to the group
                  pattern: '^[a-zA-Z0-9]+$'
                  minLength: 1
                  maxLength: 50
                  example: "user789"
              required:
                - userId
      responses:
        '201':
          description: User added to group successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 100
                    example: "User added to group successfully"
              example:
                message: "User added to group successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User already in group
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/{groupId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/groupId'
      - $ref: '#/components/parameters/userId'
    delete:
      tags: ["Groups"]
      summary: Leave group
      description: Remove the current user from a group
      operationId: leaveGroup
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Left group successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    User:
      type: object
      description: User profile information
      properties:
        id:
          type: string
          description: Unique user identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "user123"
        name:
          type: string
          description: User's display name
          minLength: 3
          maxLength: 16
          pattern: '^.*?$'
          example: "Maria"
        photoUrl:
          type: string
          format: url
          description: URL to user's profile photo
          pattern: '^https?://.+$'
          minLength: 1
          maxLength: 500
          example: "https://api.example.com/users/user123/photo"
      required:
        - id
        - name
    Group:
      type: object
      description: Group chat information
      properties:
        id:
          type: string
          description: Unique group identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "group123"
        name:
          type: string
          description: Group display name
          pattern: '^.+$'
          minLength: 1
          maxLength: 50
          example: "Family Chat"
        photoUrl:
          type: string
          format: url
          description: URL to group's photo
          pattern: '^https?://.+$'
          minLength: 1
          maxLength: 500
          example: "https://api.example.com/groups/group123/photo"
        members:
          type: array
          description: List of group members
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when group was created
          minLength: 1
          maxLength: 50
          example: "2023-12-07T10:30:00Z"
      required:
        - id
        - name
        - members
        - createdAt
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (e.g. missing fields, etc.)
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Unauthorized:
      description: The access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Forbidden:
      description: The user is not authorized to perform this operation
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |-
        Bearer authentication using user identifier.
        Use the identifier returned from the login endpoint.
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "user123"
...
    groupId:
      name: groupId
      in: path
      required: true
      description: Group identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "group123"